move checkbox utils to the folder
